Starknet's block numbers return decimal, not hex

+++ modules/consts.go
@@ -13,5 +13,5 @@ var (
 	// Health check constants
 	DefaultHCThreshold = 2
 	DefaultHCInterval  = 5
-	DefaultBlockLagLimit = 5
+	DefaultBlockLagLimit = int64(5)
 )
